arreglo de click a detalle producto

+++ app/src/main/java/com/example/myapplication/ui/home/HomeFragment.kt
@@ -103,7 +103,7 @@ class HomeFragment : Fragment() {
     // Configuración del RecyclerView
     private fun setupRecyclerView() {
         productAdapter = ProductAdapter { product ->
-            val action = HomeFragmentDirections.actionHomeFragmentToProductDetailFragment(product.codigo.toInt())
+            val action = HomeFragmentDirections.actionHomeFragmentToProductDetailFragment(product.codigo)
             findNavController().navigate(action)
         }
         binding.rvProductos.apply {
